Optimize ddvt traversal further
- Add parent pointers to QuadTree
- Add RecordStack and RecordStackIter types to ddvt
- Use pointers to iterate children of Ddvt since its always know that
  there are 4 or 0.
- Explode walk_bb_mut 6 nested logic branches into a more complex set
  that results in 4 or less tests instead of always 6.
- Fix missing Join work for Volume. Skipped uncontained after adding
  triggers to ddvt.
- Make walk functions concrete taking mutable function points instead.
- Fix clearing of removed_triggers and deeper_triggers
- Add walk_with_record_rev_mut
- Use walk_with_record_rev_mut in update_split_and_join

RecordStack and RecordStackIter use QuadTree parent pointers to
traverse parents. This removes the need to continuously rebuild a
vector of parents to traverse when moving particles to a new containing
ddvt.

Concrete walk functions instead of generic walk functions are more
likely to be inlined and as inline reduce recursive call overhead
allowing updating to be faster.