Clean up collision. Add to v2, particle, and world
- Clean up some of collision's commented code
- Normalize collision correction vector and increase its stability
- Add V2::infinity()
- Add Particle.drag and default last_position to infinity
- Adding a particle to the world with infinite last_position has
  last_position set to position
- Add Particle.friction2
- Fix World.add_particle. Wasn't respecting the id it was setting to
  where it put the particle.
- Add WorldEffect and WorldEffectRef

The collision correction vector wasn't being normalized so some
collisions could multiply its correction undesirably.

WorldEffect can modify world when its added, removed, and during each
World step. WorldEffectRef is returned for an added WorldEffect, it can
borrow the installed effect when requested with a mutable world
reference and is the only thing that can be used to remove the
installed Effect.