Extract pool logic from WorldPool
world and ddvt follow similar logic for handling threaded work that for
world was hard coded and as of this change still hard coded for ddvt.
It involves a pool of threads that take one type and result in another.
This change extracts that into a type that will create the needed
threads from a generator function which returns Consumers that take the
first type, the job, and returns options of the second type, the result.