Create an internal MiniParticle representation
MiniParticle is about a third the memory of the current Particle. It is
all thats needed to perform collision detection and correction for the
current fluid focus items. May need to expand Particle states to
consider the current Dynamic as Fluid and make a new Dynamic for
particles that need to test things like friction. These would also open
up the option for the new Dynamic to ignore collisions against specific
other Dynamics. Something we don't Fluids to be able to do as they'd
likely never use it and it added a lot of unnecessary branching since
Fluid makes up most of the simulated particles.

The benefit of the smaller memory footprint with MiniParticle is to
further avoid cache miss while testing and solving particles. The
current buffer of 256 will be a third the size too. If the Particle
compiles to take 72 bytes and MiniParticle 24 the buffer goes from ~18k
to ~6k.